fix error

+++ include/impl/iutest_port.ipp
@@ -375,7 +375,7 @@ IUTEST_IPP_INLINE ::std::string WideStringToMultiByteString(const wchar_t* wide_
 {
     if( wide_c_str == NULL ) return "";
     ::std::string str;
-    const size_t length = num < 0 ? wcslen(wide_c_str) * 2 + 1 : static_cast<size_t>(num);
+    const int length = num < 0 ? static_cast<int>(wcslen(wide_c_str) * 2 + 1) : num;
     char* mbs = new char [length];
     WideCharToMultiByte(932, 0, wide_c_str, static_cast<int>(wcslen(wide_c_str))+1, mbs, length, NULL, NULL);
     str = mbs;
